Updated to PS 5 version
The SVG's was showing a gamepad as default ,because both PS5 and PS4 versions are available. Unfortunately my SVG software renders in a strange SVG format, causing the image to bug out the program. Removed it instead. The real fix in production would be to have a graphic designer just save the playstation 4 svg in Adobe Illustrator